20Oct2024

+++ app/Http/Controllers/StudentPhonesController.php
@@ -8,7 +8,16 @@ use Illuminate\Http\Request;
 class StudentPhonesController extends Controller
 {
     public function destory(string $id){
+
+        
         $studentphone = StudentPhone::find($id);
+
+        $student = $studentphone -> student;
+        // check if profile lock 
+        if($student -> isProfileLock()){
+            return redirect()->back()->with("error","Profile Locked, please contact to admin");
+        }
+
 
         $studentphone->delete();
